Move hasNumber
