+++ linter/lints/common.go
@@ -3,6 +3,8 @@ package lints
 import (
 	"time"
 
+	"github.com/zmap/zcrypto/encoding/asn1"
+	"github.com/zmap/zcrypto/x509/pkix"
 	"github.com/zmap/zlint/v3/lint"
 )
 
@@ -21,3 +23,14 @@ const (
 var (
 	CPSV33Date = time.Date(2021, time.June, 8, 0, 0, 0, 0, time.UTC)
 )
+
+// GetExtWithOID is a helper for several of our custom lints. It returns the
+// extension with the given OID if it exists, or nil otherwise.
+func GetExtWithOID(exts []pkix.Extension, oid asn1.ObjectIdentifier) *pkix.Extension {
+	for _, ext := range exts {
+		if ext.Id.Equal(oid) {
+			return &ext
+		}
+	}
+	return nil
+}
